+++ markdown_toc_creator/exceptions.py
@@ -1,4 +1,2 @@
-class HeaderLevelNotContinuousException(Exception):
+class HeaderLevelNotContinuousError(Exception):
     """Exception when header levels are not continuous"""
-
-    pass